Отображение опций и пункта назначения в форме редактирования объединены в один элемент. Отредактировано отображение формы с пустыми или отсутствующими данными
