+++ src/mock/trip-points.js
@@ -63,7 +63,7 @@ const mockTripPoints = [
   },
   {
     'id': 6,
-    'type': TRIP_TYPES[6],
+    'type': TRIP_TYPES[7],
     'destination': getRandomArrayElement(mockDestinations).id,
     'dateFrom': new Date('11/12/24 22:00').toISOString(),
     'dateTo': new Date('11/15/24 22:40').toISOString(),
